requirements/production.txt
gunicorn==19.9.0
asgiref==3.2.10
Django==3.1
django-redis==4.12.1
django-redis-cache==2.1.1
djangorestframework==3.11.1
psycopg2==2.8.5
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
